+++ CHANGELOG.md
@@ -2,6 +2,8 @@ v1.1.0 (in development)
 -----------------------
 - Use pydantic internally for parsing & validating version databases
 - `eol_date()` and `is_eol()` now accept major and micro versions
+- The `show` command now outputs the EOL date and EOL state for all CPython
+  version levels
 
 v1.0.0 (2021-11-04)
 -------------------